requirements.txt
asgiref==3.2.3
backports.shutil-get-terminal-size==1.0.0
cffi==1.13.2
cryptography==2.8
decorator==4.0.10
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-import-export==2.0.1
enum34==1.1.6
et-xmlfile==1.0.1
ipython-genutils==0.1.0
jdcal==1.4.1
MarkupPy==1.14
more-itertools==8.0.2
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycparser==2.19
pycrypto==2.6.1
Pygments==2.5.2
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.3.0
tablib==0.14.0
traitlets==4.3.3
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
